requirements.txt
numpy
torch==1.13
torchvision
torchaudio
pytorch-lightning
pydegensac==0.1.2
opencv_python
pickle5
hydra-core==1.1.1
omegaconf==2.1.2
pycocotools
wandb
rich
transforms3d
natsort
pymatting
accelerate
fire
pillow
scikit-image
tqdm
scikit-learn
SimpleCRF
albumentations
loguru
h5py